Move deprecation utilities to private namepsace
